+++ VSP.Tests/Import/ImportWizardViewModelTests.cs
@@ -4,7 +4,7 @@ using VSP.Device.Import.Execution;
 using VSP.Device.Import.Preview;
 using VSP.Device.Import.Validation;
 using VSP.Device.Interfaces;
-using VSP.Domain.Entities;
+using EntityCamera = VSP.Domain.Entities.Camera;
 using VSP.UI.Helpers;
 using VSP.UI.ViewModels;
 using Xunit;
@@ -335,24 +335,24 @@ public class ImportWizardViewModelTests : IDisposable
 
     private sealed class FakeCameraRepository : ICameraRepository
     {
-        public List<Camera> Cameras { get; } = [];
+        public List<EntityCamera> Cameras { get; } = [];
 
-        public IEnumerable<Camera> GetAll()
+        public IEnumerable<EntityCamera> GetAll()
         {
             return Cameras;
         }
 
-        public Camera? GetById(Guid id)
+        public EntityCamera? GetById(Guid id)
         {
             return Cameras.FirstOrDefault(camera => camera.Id == id);
         }
 
-        public void Add(Camera camera)
+        public void Add(EntityCamera camera)
         {
             Cameras.Add(camera);
         }
 
-        public void Update(Camera camera)
+        public void Update(EntityCamera camera)
         {
             throw new NotSupportedException();
         }